django_test/requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
pylint==2.16.3
djangorestframework==3.14.0
PyJWT==2.6.0